requirements.txt
bcrypt
pyjwt
flask
flask_restful
gevent
pickledb
uuid
